Task mark as completed

+++ pages/api/tasks/[id].ts
@@ -5,10 +5,12 @@ export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse
 ) {
-  const data = await axios.delete(
-    `${process.env.BASE_URL}/tasks/${req.query.id}`
-  );
-  const [response] = await Promise.all([data]);
+  if (req.method === 'DELETE') {
+    const data = await axios.delete(
+      `${process.env.BASE_URL}/tasks/${req.query.id}`
+    );
+    const [response] = await Promise.all([data]);
 
-  res.status(200).json(response.data);
+    res.status(200).json(response.data);
+  }
 }
